Deprecate using `#serialize` with boolean argument
Update `Axlsx::Package#serialize` to accept the second argument as a
boolean (being deprecated) or an options hash.

In order to transition toward using keyword arguments for
`Axlsx::Package#serialize`, change the documented method signature to an
options hash, while still parsing the second argument as `confirm_valid`
if a boolean is provided (in which case we also warn the user that a
boolean argument is deprecated).