Completely hide chart titles if blank
Before this change, a chart with a blank title would show up in Excel with a default title like “Chart title” or “Diagrammtitel”.

It was not possible to completely hide a chart title: Setting it to an empty space would override Excel’s default, but the empty title did still occupy vertical space.

Per OOXML standard the `title` element is optional, so omitting it if the title is blank looks like the right thing to do and seems to work correctly in Excel.