Prevent Excel from crashing when multiple data columns are added to PivotTable
Excel will crash unless both `colFields` and `colItems` are declared in the PivotTable xml.

First `field` of `colFields` needs to be set to -2 to tell Excel that the value columns comes from individual columns and not groups.
I do not know why `colItems` are needed, but when set, the Excel document opens without crashing.