Introduce Axlsx.skip_validations in order to improve speed / memory
The `Axlsx#validate_*` methods are used all over the place to check
the type / value of attributes. This can be useful in some cases, but
in many situation that checks are not necessary. In some ways this is
similar to `Axlsx.trust_input` where sometimes we just want to go
fast.

This commit introduced `Axlsx.skip_validations` which by passing
all the checks. This is implemented as a simple short-circuit in
each of the validate methods which already has ton of impact. In
the future, it could be useful to find another implemenation as
to avoid all those method invocations.