Derive SimpleTypedList from Array
Currently the SimpleTypedList is wrapping `@list` Array and re-implementing
certain methods to add type / value checking and delegating all the
non-destrutive methods to the internal @list.

Most everything that ends up being serialized to the excel file is derived
from SimpleTypedList (for example, a Row is a SimpleTypedList of Cells). This
means that many of the delegated methods are hit over and over again especially
`[]`. This extra delegation really adds up.

We can avoid the delegation by having SimpleTypedList extend Array and still
remove any un-safe methods.

It is possible to re-introduce an unsafe method in a subclass of SimpleTypedList
if needed. Here is one way,
```ruby
class MyList < SimpleTypedList
  def clear
    @clear_method ||= Array.instance_method(:clear).bind(self)
    @clear_method.call
  end
end
```