Only define @escape_formulas in cell if it is different from worksheet
The benchmarks showed that validate_boolean is called 200_005 times
and almost all of those are to validate escape_formulas passed into
cell.

In this commit the worksheet does not pass in its escape_formulas
value, avoiding validate_boolean, and instead the cell asks the
worksheet for value when needed. Now validate_boolean is called 5
times in benchmarks.