Fix Style/AndOr offenses
From Ruby Style Guide:

> Do not use `and` and `or` in boolean context - and and or are control
flow operators and should be used as such. They have very low
precedence, and can be used as a short form of specifying flow sequences
like "evaluate expression 1, and only if it is not successful
(returned `nil`), evaluate expression 2". This is especially useful for
raising errors or early return without breaking the reading flow.

Also:

- Remove redundant use of self
- Use attribute reader instead of accessing instance variable