Avoid unnecessary calls to `parse_options`
`Cell#initialize` is one of the most used methods in the library, so
a small improvement here will make a lot of difference on the overall
performance.

The impact of this change is negligible when `options` has elements
but is substantial when it is empty.

```
Comparison (empty options):
        parse unless: 11636650.6 i/s
               parse:  8109825.4 i/s - 1.43x  (± 0.00) slower

Comparison (1 option):
               parse:  3548037.5 i/s
        parse unless:  3459029.7 i/s - same-ish: difference falls within error
```