Add `uri_parser` method for RFC2396 compatibility
This commit introduces a new method `Axlsx#uri_parser` to handle
URI parsing in a way that's compatible with both newer and older
versions of Ruby's URI library. It checks for the presence of
`URI::RFC2396_PARSER` and falls back to `URI::DEFAULT_PARSER` if
not available, ensuring consistent behavior across different Ruby
versions.

Close #397

Ref: ruby/uri#114